refactor(schema): split lead-scoring schema into schemes/lead_scoring/ [LTV-Pg.2]
Completes the M2 physical reorg (LTV-Pg). Pulls the lead-scoring-specific
schema definitions out of the shared `schema/` package, leaving only genuine
cross-scheme primitives behind.

New files (schemes/lead_scoring/):
- entities.py   — ContactRow … SubscriptionRow + ALL_ROW_TYPES / TABLE_NAMES
- relationships.py — ALL_CONSTRAINTS
- features.py  — LEAD_SNAPSHOT_FEATURES + redacted_columns_for
- tasks.py     — CONVERTED_WITHIN_90_DAYS + task_manifest_for_config

Shared schema/ after the split (only primitives remain):
- entities.py      — EntityRowProtocol, make_empty_dataframe, AccountRow
- features.py      — FeatureSpec
- relationships.py — FKConstraint, FKViolationError, validate_fk
- tasks.py         — SplitSpec, TaskManifest

All callers updated (36 files): multi-line from-import blocks rewritten via
perl; 4 mixed imports (tests/exposure/test_redaction.py, tests/schema/
test_relationships.py, tests/render/test_render.py, tests/narrative/
test_dataset_card.py) fixed manually where FeatureSpec/validate_fk/TaskManifest
stay shared but co-imported with moved symbols.

tests/schemes/test_module_layout.py: 3 new tests for Pg.2 — primitives-stay,
scheme-specifics-in-scheme, removed-from-shared-schema.

CHANGELOG, CLAUDE.md (canonical layout), roadmap (Pg.2 ✓), agent-plan updated.

Verified byte-identical to pre-reorg main (14/14 files); full suite
1537 passed / 51 skipped; ruff + mypy clean (96 source files).
