+++ .agent-plan.md
@@ -48,8 +48,9 @@ snapshots/relational_snapshot_safe/tasks moved + relational.py split so the
 shared write_relational_tables stays in the envelope) merged (#110). `LTV-Pg`
 (schema reorg) split: Pg.1 (scaffold `schemes/lifecycle/` — lifecycle rows +
 registries + LIFECYCLE_CONSTRAINTS moved there, stub LifecycleScheme registered)
-opened as **#111**; Pg.2 (split the lead-scoring schema out of shared `schema/`)
-pending. All M2 moves byte-identical. Sibling `leadforge-datasets-private`
+merged (#111); Pg.2 (split lead-scoring schema: entity rows/ALL_ROW_TYPES/
+ALL_CONSTRAINTS/LEAD_SNAPSHOT_FEATURES/CONVERTED_WITHIN_90_DAYS moved to
+`schemes/lead_scoring/`; shared primitives stay in `schema/`) opened as **#NNN**. All M2 moves byte-identical. Sibling `leadforge-datasets-private`
 consumes bundle files, not internals — no lockstep update needed (heads-up
 issue #8). Next: `LTV-Pg.2`, then `LTV-Pc` (pLTV feature/task specs, authored in
 `schemes/lifecycle/`), then `LTV-M3` (lifecycle population).
