+++ tests/schemes/test_module_layout.py
@@ -27,6 +27,7 @@ _MOVED = [
         "leadforge.schemes.lead_scoring.render.relational_snapshot_safe",
     ),
     ("leadforge.render.tasks", "leadforge.schemes.lead_scoring.render.tasks"),
+
 ]
 
 
@@ -68,6 +69,44 @@ def test_relational_split_to_dataframes_moved_to_scheme() -> None:
     # The ambiguous flat `leadforge.render.relational` module is gone.
     with pytest.raises(ModuleNotFoundError):
         importlib.import_module("leadforge.render.relational")
+
+
+def test_schema_split_primitives_stay_in_schema() -> None:
+    # LTV-Pg.2: shared primitives kept in schema/, not moved with the rows.
+    from leadforge.schema.entities import (  # noqa: F401
+        AccountRow,
+        EntityRowProtocol,
+        make_empty_dataframe,
+    )
+    from leadforge.schema.features import FeatureSpec  # noqa: F401
+    from leadforge.schema.relationships import FKConstraint, validate_fk  # noqa: F401
+    from leadforge.schema.tasks import SplitSpec, TaskManifest  # noqa: F401
+
+
+def test_schema_split_lead_scoring_specifics_in_scheme() -> None:
+    # LTV-Pg.2: lead-scoring-specific symbols live in the scheme package.
+    from leadforge.schemes.lead_scoring.entities import (  # noqa: F401
+        ALL_ROW_TYPES,
+        ContactRow,
+        LeadRow,
+    )
+    from leadforge.schemes.lead_scoring.features import LEAD_SNAPSHOT_FEATURES  # noqa: F401
+    from leadforge.schemes.lead_scoring.relationships import ALL_CONSTRAINTS  # noqa: F401
+    from leadforge.schemes.lead_scoring.tasks import CONVERTED_WITHIN_90_DAYS  # noqa: F401
+
+
+def test_schema_split_lead_scoring_removed_from_shared_schema() -> None:
+    # LTV-Pg.2: moved symbols are gone from the shared schema namespace.
+    import leadforge.schema.entities as shared_entities
+    import leadforge.schema.features as shared_features
+    import leadforge.schema.relationships as shared_relationships
+    import leadforge.schema.tasks as shared_tasks
+
+    assert not hasattr(shared_entities, "LeadRow")
+    assert not hasattr(shared_entities, "ALL_ROW_TYPES")
+    assert not hasattr(shared_features, "LEAD_SNAPSHOT_FEATURES")
+    assert not hasattr(shared_relationships, "ALL_CONSTRAINTS")
+    assert not hasattr(shared_tasks, "CONVERTED_WITHIN_90_DAYS")
 
 
 def test_public_api_unchanged_by_the_move() -> None:
